[IMP] mail: use thread displayName instead of name
More often than not these 2 fields have the same value, but when they are
different it is `displayName` that holds the expected value (eg. for a DM,
`displayName` would be the custom name or the name of the correspondent).

This commit updates them to avoid confusion and having bad examples in the code.

closes odoo/odoo#73659

Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>