[FIX] web_editor: properly handle media "text tools" options
Since their convertion to "snippet" options with [1], the media options
were not handling previews correctly anymore, but also were recoding
their entire logic for each of the options instead of using generic
methods like selectClass or selectStyle.

One example of a bug was that previewing the 25% width option on an
image, not enabling it and saving the editor forced the image to a fixed
px width (its original one). Changing that image to another would keep
that wrong fixed width.

Note: this reveals that selectClass and selectStyle need more parameters
to handle all of this correctly. Most was added as a specific override
for these features but it could be handled even more generically.

[1]: https://github.com/odoo/odoo/commit/e3baa81b0a27b0757633c89ade9b77d02671cb91

Related to task-2578242
task-2577848

closes odoo/odoo#73734

X-original-commit: dafc8dc3e52918653aa9852ded85fce5bef1eab6
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>