[FIX] website: avoid traceback when clicking on stat buttons
Prior to this commit, if we go to visitors form view and clicked on
stat buttons like offline/connected or visits, a traceback was
thrown because the buttons are meant for showing stat only, they
aren't linked to real actions / methods. It used to work before, but
the recent traceback is result of more strict checks from the web-client
re-write. See the original commit[1] for more information.

This commit sets 'disabled' attribute on the buttons, and keeps the action
from being performed for display-only stat buttons.

commit[1] - https://github.com/odoo/odoo/commit/6940629267319756f9f3ebd4faf2620372a2905f

TaskID-2593552

closes odoo/odoo#73813

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>