[FIX] mail_plugin: fix an error during the authentication
Bug
===
Since 3f80ecc9decc9908694fe0dd3e4e6c064f7cf12c we wanted to use the
Odoo redirection function everywhere.

But this function work only if we redirect the user to the same site.

The mail plugin need to redirect outside of Odoo and so we can not use
this function.

closes odoo/odoo#73698

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>