[FIX] web: fallback for cache_hashes when http_routing is not installed
Previously, loading the login page with no modules installed would
result in a crash, because the translation service assumed that there
was always a cache_hashes key in the session info. This is not the case
when the http_routing module is not installed.

This module fixes that by adding a fallback for those cache_hashes in
the translation service.

closes odoo/odoo#73874

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>