[IMP] account: payment receipt improvements
The current template used to render the payment receipt
lack of details, which cause it to sometimes not make
much sense.

Eg.
    Invoice of  3000
    Reversal of 1000
    Payment of  2000
The receipt won't display information about the reversal,
showing something akin to "3000-2000 = 0"

This will improve the template to give more detail to the
user and avoid possible incomprehension.

Task id #2545590

closes odoo/odoo#73974

Signed-off-by: William André (wan) <wan@odoo.com>