[FIX] l10n_es: Make non deductible taxes to propagate analytic
When you are using non deductible taxes, the tax amount is handled as
an expense increase, using the same expense account than the invoice
line. But not having the analytic option, this expense increase is not
classified in the same analytic account.

Thus, a sane default for the initial chart of accounts is to set it to
True.

TT31062

closes odoo/odoo#74111

X-original-commit: bb795f3cacef5a47a8e41cdfc31b85bb21582016
Signed-off-by: William André (wan) <wan@odoo.com>