[IMP] test_website: add tests for image upload progress bar
Progress bar on image upload was introduced with #65828 in saas-14.3.

Since, it has been broken in saas-14.3 with 3f23da4 which broke the flow for
image unrecognized by PIL, see the other commit of this PR.

It was also completely broken in master (saas-14.5) with the wowl refactoring,
where no progress bar were shown at all. It would show empty toastr.
This was fixed with #74027

This commit introduces a complete testing suite for that progress bar:
  - Single image upload
  - Multi image upload
  - Success upload
  - Unsupported error
  - Unrecognized error

task-2607393
Closes #74122

closes odoo/odoo#74454

X-original-commit: 696bbceb8bcbe83a1ad45d88b10045b8b2a06443
Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>