[FIX] point_of_sale: Don't fail when manually running scheduler
When running the scheduler manually with a user who has no read access on pos.session (and point_of_sale is installed),
the run fails silently for the user, and shows an error in the logs.
"_alert_old_session" runs with the user's rights, who might have POS access.

This commits runs the problematic search with sudo to work around this
issue.

closes odoo/odoo#74426

X-original-commit: abd1966d094ca3d85ff41d33fe0169de6b5f1add
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>